sched/fair: upadte adjust_cpus_for_packing()
Two changes:

- Prefer idle cpu over an active CPU if the idle CPU is not
in a deep c-state than WFI.

- if target_cpus_count > 1 means, we have found two active
CPUs. so we assume packing on this active cpu does not raise
the frequency any more. since we select the target_cpu as the
least loaded cpu, we can go ahead and check if it can accmmodate
the task without raising the frequency or not. so remove that
limitation as well.

Change-Id: I0c226e977d28d8948ec9f09cf6a19f35420b3fd3
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: RuRuTiaSaMa <1009087450@qq.com>