mailbox: msm_qmp: Remove IRQF_NO_SUSPEND
The qmp interrupt is a hardware wake capable interrupt. Using the
IRQF_NO_SUSPEND flag still allows this interrupt to resume the CPUs
because the interrupt is left unmasked on the gic. Remove this flag
and use enable_irq_wake so the wakeup interrupt is properly accounted
for.

Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Ananth Raghavan Subramanian <sananth@codeaurora.org>
Signed-off-by: Kelly Rossmoyer <krossmo@google.com>
Change-Id: I7e55b9feff43da6bda5d1eca12bf92f98be1c9f7
Bug: 123377615
Bug: 131260677
Signed-off-by: alk3pInjection <webmaster@raspii.tech>
(cherry picked from commit 4abdf21a05b714238424b0705a257035a191ca6f)
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>