soc: qcom: scm: Don't allocate memory dynamically for extra arg
On RT kernels, page allocations cannot be used from atomic context, and
scm_call2_atomic() is used from atomic context on RT, so the dynamic memory
allocation for the extra arg needs to go. As it turns out, the stack depth
is not too bad when reaching scm_call2_atomic() and __scm_call2 from the
most stack-hungry call chains: about half the stack is free in the worst
case when called from kernfs_fop_write() which uses a 4K stack buffer.

Since there's stack space to spare, move the page-aligned extra arg buffer
to the stack. This burns just slightly above 4K in stack space, but it's
safe.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>